Additions to MiniDST.xml (#151)
* Add CPID calibration for ILD MC production of 2020, one based on training with single particles and one based on 2-fermion-Z-hadronic events.
This includes the weight files and the reference files which point to the weight files.
In addition, the CPID steering is added to the MarlinRecoMiniDST.xml.
Paths are relative and only work if executed from the StandardConfig/production folder!

* additional minor changes in response to PR review

* Add CPID calibration for ILD MC production of 2020, one based on training with single particles and one based on 2-fermion-Z-hadronic events.
This includes the weight files and the reference files which point to the weight files.
In addition, the CPID steering is added to the MarlinRecoMiniDST.xml.
Paths are relative and only work if executed from the StandardConfig/production folder!

* additional minor changes in response to PR review

* add: change to MiniDST steering file to include new CPID p > 1 GeV weight files

* Additions to the MiniDST.xml

* typo

* Put constant into quotes

Marlin XML parsing doesn't care, but other xml parsers might

---------

Co-authored-by: Thomas Madlener <thomas.madlener@desy.de>